chore: move set background listeners to pushNotification.js
