+++ index.js
@@ -8,55 +8,16 @@
 import './i18nInit';
 import { AppRegistry } from 'react-native';
 import { setJSExceptionHandler } from 'react-native-exception-handler';
-import messaging from '@react-native-firebase/messaging';
-import notifee, { EventType } from '@notifee/react-native';
 import App from './src/App';
 import { name as appName } from './app.json';
 import { errorHandler } from './src/errorHandler';
-import { messageHandler, setInitialNotificationData } from './src/sagas/pushNotification';
+import {
+  setNotifeeBackgroundListener,
+  setBackgroundMessageListener
+} from './src/sagas/pushNotification';
 
-/**
- * This function is called when the app is in background or **quit** state
- * and a push notification from **notifee** is received. It also handles
- * notification events like dismiss and action press.
- *
- * @see https://notifee.app/react-native/reference/eventtype
- *
- * In this project the Notifee is used to interprete the data message received
- * from firebase and show the notification to the user.
- */
-const setNotifeeBackgroundListener = () => {
-  try {
-    notifee.onBackgroundEvent(async ({ type, detail }) => {
-      const { notification } = detail;
-
-      if (type === EventType.DISMISSED) {
-        await notifee.cancelNotification(notification.id);
-        return;
-      }
-
-      if (type === EventType.PRESS || type === EventType.ACTION_PRESS) {
-        setInitialNotificationData(notification);
-        await notifee.cancelNotification(notification.id);
-      }
-    });
-  } catch (error) {
-    console.error('Error setting notifee background listener.', error);
-  }
-};
+// Set the background listeners for push notifications
 setNotifeeBackgroundListener();
-
-/**
- * Install the listener to handle push notifications when the application is in background or quit.
- */
-const setBackgroundMessageListener = () => {
-  const onBackgroundMessage = async (message) => messageHandler(message);
-  try {
-    messaging().setBackgroundMessageHandler(onBackgroundMessage);
-  } catch (error) {
-    console.error('Error setting background message listener.', error);
-  }
-};
 setBackgroundMessageListener();
 
 // - errorHandler is the callback
