+++ src/sagas/pushNotification.js
@@ -19,7 +19,7 @@ import {
   fork,
 } from 'redux-saga/effects';
 import messaging from '@react-native-firebase/messaging';
-import notifee from '@notifee/react-native';
+import notifee, { EventType } from '@notifee/react-native';
 import { msgid, ngettext, t } from 'ttag';
 import { Platform } from 'react-native';
 import {
@@ -139,6 +139,48 @@ const localization = {
       message = t`New transaction received`;
     }
     return message;
+  }
+};
+
+/**
+ * This function is called when the app is in background or **quit** state
+ * and a push notification from **notifee** is received. It also handles
+ * notification events like dismiss and action press.
+ *
+ * @see https://notifee.app/react-native/reference/eventtype
+ *
+ * In this project the Notifee is used to interprete the data message received
+ * from firebase and show the notification to the user.
+ */
+export const setNotifeeBackgroundListener = () => {
+  try {
+    notifee.onBackgroundEvent(async ({ type, detail }) => {
+      const { notification } = detail;
+
+      if (type === EventType.DISMISSED) {
+        await notifee.cancelNotification(notification.id);
+        return;
+      }
+
+      if (type === EventType.PRESS || type === EventType.ACTION_PRESS) {
+        setInitialNotificationData(notification);
+        await notifee.cancelNotification(notification.id);
+      }
+    });
+  } catch (error) {
+    console.error('Error setting notifee background listener.', error);
+  }
+};
+
+/**
+ * Install the listener to handle push notifications when the application is in background or quit.
+ */
+export const setBackgroundMessageListener = () => {
+  const onBackgroundMessage = async (message) => messageHandler(message);
+  try {
+    messaging().setBackgroundMessageHandler(onBackgroundMessage);
+  } catch (error) {
+    console.error('Error setting background message listener.', error);
   }
 };
